restructure workout and routine components by feature
