+++ src/app/(main)/workout/_components/WorkoutPlaceholder.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import RoutineList from "@/app/(main)/routines/_components/RoutineList";
+import RoutineList from "@/app/(main)/routines/_components/routineList/RoutineList";
 import {
   routineDetailService,
   workoutDetailAdapter,
