+++ src/app/(main)/workout/_components/expandedView/ExpandedWorkoutItem.tsx
